refactor(web): extract useUrlSync hook from page.tsx
Move syncUrlToState, the initial-sync effect, and the popstate listener into
lib/hooks/useUrlSync.ts. The hook takes only the state/setters it needs,
decoupling URL syncing from the page component.
